app_pc/requirements_for_app .txt
# requirements.txt
# Core dependencies với versions cụ thể

# Xử lý video và ảnh
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
Pillow==10.1.0

# Machine Learning
scikit-learn==1.3.2
lightgbm==4.1.0

# Deep Learning (optional, CPU-first)
torch==2.1.0
torchvision==0.16.0

# Data processing
pandas==2.1.4
PyYAML==6.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# PDF generation
reportlab==4.0.7

# GUI
PySide6==6.6.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
joblib==1.3.2